requirements.txt
einops
gdown
imageio
ipykernel
matplotlib
numpy
pandas
PyYAML
scikit-image
scikit-learn
scipy
tensorboard
tensorboardX
torch-fidelity
torchmetrics
tornado
tqdm